Refactor rust
